requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
lxml==4.4.2
numpy==1.18.1
pandas==0.25.3
pyhunter==1.3
pyperclip==1.7.0
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.7
